🚚 rename: 상수 common.js 생성 및 SPACE common.js로 옮기기 (#197)
